commit backend filewrite progress

+++ services/news_fetcher.py
@@ -13,13 +13,26 @@ bbcRssDict = rss_sources.getBbcRss()
 
 def getNews(rssDict):
 	for key, value in rssDict.items():
-		fileName = "./data/" + key + ".xml"
-		os.remove(fileName)
-		fileObj = io.open(fileName, 'w', encoding='utf8')
+		fileName = "./data/" + key + "-write.json"
+		if os.path.exists(fileName):
+			os.remove(fileName)
+		fileObj = io.open(fileName, 'wb')
 		feed = feedparser.parse(value)
+		feedData = []
 		for post in feed.entries:
 			summary = SummarizeUrl(post.link)
-			fileObj.write(post.title + ": " + json.dumps(summary) + "\n")
+			summaryJson = json.dumps(summary)
+
+			dataHolder = {}
+			dataHolder['summary'] = summaryJson
+			dataHolder['title'] = post.title
+			dataHolder['pstUrl'] = post.link
+
+			feedData.append(json.dumps(dataHolder))
+		feedDataJson = json.dumps(feedData, sort_keys=True, indent=4)
+		
+		fileObj.write(feedDataJson)
+		print('wrote file: %s' % fileName)
 		fileObj.close
